requirements.txt
Flask==3.0.0
numpy==1.26.1
opencv-python==4.8.1.78
pyserial==3.5